Add machine trap handlers
Since the physical address of the UART is stored as a constant in memory
and is not a symbol placed at a particular address, which would be
computed as a PC-relative address and thus automatically account for VM,
virtual address translation must be bypassed in order to write there.
The way to work around this is to execute the store in machine mode,
which bypasses VM. So, to do that, an attempt to perform a write system
call from S-mode will make another ecall to get to M-mode.